initial structure
This is my proposal for the project structure. Not all components are
fully implemented but it should be enough to show my general idea. The
storage and service package provide interfaces to allow easier testing
and to be able to provide different implementations e.g. we could have a
postgres and a sqlite storage implementation. The api package currently
implements a REST api but this can be replace with a graph api. I don't
have any hard opinions about that.

I'm also open for other suggestions or any feedback in general. My main
goal for the structure is to have a clean seperation of
responsabilities.

This minimal PoC can be tested by starting the server like so (you have
to be in the project root folder):
```
$ go run cmd/cuttables/main.go server
```
Then you can send a curl request to list all cuttables:
```
$ curl http://localhost:8080/api/v1/cuttables
```